requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
Pillow>=9.5.0
reportlab>=4.0.0
python-decouple>=3.8
whitenoise>=6.5.0
